+++ CHANGES.md
@@ -17,6 +17,7 @@ Changes in 12.4
 - Fix PHP warning undefined variable $date_select when no calendars in Eligibility/TeacherCompletion.php
 - Update French & Spanish translations in rosariosis.po
 - Security: add `.htaccess` file to block direct access to potentially sensitive files in .htaccess
+- CSS WPadmin theme fix password strength bars in stylesheet.css
 
 Changes in 12.3.2
 -----------------